+++ src/agat/agat_sp_clipn_seqextremities_and_fixcoordinates/test.sh
@@ -1,0 +1,37 @@
+#!/bin/bash
+
+## VIASH START
+## VIASH END
+
+test_dir="${meta_resources_dir}/test_data"
+out_dir="${meta_resources_dir}/out_data"
+
+echo "> Run $meta_name with test data"
+"$meta_executable" \
+  --gff "$test_dir/input.gff" \
+  --fasta "$test_dir/input.fasta" \
+  --output_fasta "$out_dir/output.fasta" \
+  --output_gff "$out_dir/output.gff"
+
+echo ">> Checking output"
+[ ! -f "$out_dir/output.gff" ] && echo "Output file output.gff does not exist" && exit 1
+[ ! -f "$out_dir/output.fasta" ] && echo "Output file output.fasta does not exist" && exit 1
+
+echo ">> Check if output is empty"
+[ ! -s "$out_dir/output.gff" ] && echo "Output file output.gff is empty" && exit 1
+[ ! -s "$out_dir/output.fasta" ] && echo "Output file output.fasta is empty" && exit 1
+
+echo ">> Check if output matches expected output"
+diff "$out_dir/output.gff" "$test_dir/output.gff"
+if [ $? -ne 0 ]; then
+  echo "Output file output.gff does not match expected output"
+  exit 1
+fi
+
+diff "$out_dir/output.fasta" "$test_dir/output.fasta"
+if [ $? -ne 0 ]; then
+  echo "Output file output.fasta does not match expected output"
+  exit 1
+fi
+
+echo "> Test successful"